+++ app/components/helpers/imagepicker.js
@@ -1,23 +1,24 @@
 import { Platform, DeviceEventEmitter, NativeModules } from 'react-native';
 import ImagePicker from 'react-native-image-picker';
 import moment from 'moment';
-import { tx } from '../utils/translator';
 import { fileHelpers } from '../../lib/icebear';
 
 const { FilePickerManager } = NativeModules;
 
 let lastCall = null;
 
-const ANDROID_PICK_ACTION = 'android-pick';
-
-const showImagePicker = async options => new Promise(resolve =>
-    ImagePicker.showImagePicker(options, resolve));
+const launchGallery = async options => new Promise(resolve =>
+    ImagePicker.launchImageLibrary(options, resolve));
 
 const launchCamera = async options => new Promise(resolve =>
     ImagePicker.launchCamera(options, resolve));
 
 const showFilePicker = async () => new Promise(resolve =>
     FilePickerManager.showFilePicker(null, resolve));
+
+function waitForPermissions() {
+    return new Promise(resolve => { lastCall = resolve; });
+}
 
 function normalizeUri(response) {
     const { uri } = response;
@@ -27,68 +28,55 @@ function normalizeUri(response) {
     return uri;
 }
 
-function processResponse(response, imageCallback) {
+async function processResponse(functor, params) {
+    let response = await functor(params);
+    if (response.error) {
+        console.log('imagepicker.js: ', response.error);
+        throw new Error(response.error);
+    }
+    if (response.didRequestPermission) {
+        console.log('imagepicker.js: permissions requested');
+        await waitForPermissions();
+        response = await functor();
+    }
     if (!response.path && response.uri) {
         response.path = response.uri;
     }
-    if (response.isAndroidCamera) {
+    if (params.isCamera) {
         const ext = fileHelpers.getFileExtension(response.path);
         response.fileName = `${moment(Date.now()).format('llll')}.${ext}`;
     }
-    imageCallback(normalizeUri(response), response.fileName, response);
+    const { fileName, path, uri } = response;
+    const normalizedFileName = fileHelpers.getFileName(fileName || path || uri);
+    const ext = fileHelpers.getFileExtension(normalizedFileName);
+    return { url: normalizeUri(response), fileName: normalizedFileName, ext, response };
 }
 
+const cameraSettings = {
+    isCamera: true,
+    noData: true,
+    storageOptions: {
+        skipBackup: true,
+        waitUntilSaved: true
+    }
+};
+
+const gallerySettings = { noData: true };
+
 export default {
-    showImagePicker,
+    launchGallery,
     showFilePicker,
 
-    async show(_customButtons, imageCallback, customCallback) {
-        const customButtons = _customButtons || [];
-        const options = {
-            customButtons,
-            noData: true,
-            storageOptions: {
-                skipBackup: true,
-                waitUntilSaved: true
-            }
-        };
+    getImageFromCamera() {
+        return processResponse(launchCamera, cameraSettings);
+    },
 
-        if (Platform.OS === 'android') {
-            options.customButtons.push({ name: ANDROID_PICK_ACTION, title: tx('button_pickFromLibrary') });
-            options.chooseFromLibraryButtonTitle = null;
-        }
+    getImageFromGallery() {
+        return processResponse(launchGallery, gallerySettings);
+    },
 
-        let response = await showImagePicker(options);
-        console.log(`imagepicker.js: got response`);
-        console.debug(response);
-        // user selected camera and needs to confirm permissions
-        // TODO: check for iOS
-        if (response.didRequestPermission && response.option === 'launchCamera') {
-            console.log('imagepicker.js: requested permission');
-            lastCall = async () => processResponse(await launchCamera(options), imageCallback);
-        // user cancelled
-        } else if (response.didCancel) {
-            console.log('imagepicker.js: user cancelled image picker');
-        // user error
-        } else if (response.error) {
-            console.log('imagepicker.js: ', response.error);
-        // user selected custom button in the action sheet
-        } else if (customCallback && response.customButton && response.customButton !== ANDROID_PICK_ACTION) {
-            console.log('imagepicker.js:', response.customButton);
-            customCallback(response.customButton);
-        // user is on Android and uses custom file picker therefore
-        } else if (response.customButton === ANDROID_PICK_ACTION) {
-            lastCall = async () => {
-                response = await showFilePicker();
-                if (response.didRequestPermission) return;
-                lastCall = null;
-                imageCallback(normalizeUri(response), response.fileName, response);
-            };
-            lastCall();
-        // user selected camera or gallery and permissions are intact
-        } else {
-            processResponse(response, imageCallback);
-        }
+    getImageFromAndroidFilePicker() {
+        return processResponse(showFilePicker);
     }
 };