+++ app/components/layout/back-icon.js
@@ -3,5 +3,7 @@ import HeaderIconBase from './header-icon-base';
 
 export default class BackIcon extends HeaderIconBase {
     icon = 'arrow-back';
-    action = () => routerMain.back();
+    action = () => {
+        return this.props.action ? this.props.action() : routerMain.back();
+    }
 }