+++ app/styles/vars.js
@@ -1,15 +1,29 @@
 import { Platform, Dimensions } from 'react-native';
 import branding from './branding';
 
+const { width, height } = Dimensions.get('window');
+
+function isIphoneX() {
+    const { OS, isPad, isTVOS } = Platform;
+    const dim = 812;
+    return (
+        OS === 'ios' && !isPad && !isTVOS && (width === dim || height === dim)
+    );
+}
+
+const iPhoneXTop = isIphoneX() ? 16 : 0;
+const iPhoneXBottom = iPhoneXTop;
+
 const { bg, bgGradient, tabsBg, tabsFg } = branding;
 
-const statusBarHeight = Platform.OS === 'android' ? 0 : 10;
+const statusBarHeight = (Platform.OS === 'android' ? 0 : 10) + iPhoneXTop;
 const layoutPaddingTop = statusBarHeight * 2;
 
 const r = 40;
 const retentionOffset = { top: r, left: r, bottom: r, right: r };
 
-const { height } = Dimensions.get('window');
+const tabCellHeight = 56;
+
 // pixel ratio should be factored into scaleDim somehow: const pixRatio = PixelRatio.get();
 const defaultHeight = 667;
 // scaleDim takes a size value and returns one that is adjusted to the height of the device as it compares to an iPhone 6
@@ -60,8 +74,7 @@ const vars = {
     footerMarginX: 24,
     statusBarHeight,
     layoutPaddingTop,
-    headerHeight: 80,
-    tabsHeight: 56,
+    headerHeight: 80 + iPhoneXTop,
     headerSpacing: 56 + layoutPaddingTop,
     iconSize: 24,
     iconSizeSmall: 16,
@@ -104,6 +117,10 @@ const vars = {
     fabRight: 16,
     fabBottom: 32,
     retentionOffset,
+    iPhoneXBottom,
+    iPhoneXTop,
+    tabCellHeight,
+    tabsHeight: tabCellHeight + iPhoneXBottom,
     spacing: {
         small: {
             mini: scaleDim(2),